pyproject.toml
[tool.poetry]
name = "dds"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
dotenv = "^0.9.9"
pillow = "^11.3.0"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.16.0"
django-filter = "^25.1"


[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
